move to spec instead of nodespec
